+++ ouroboros-network/src/Ouroboros/Network/Diffusion/P2P.hs
@@ -103,17 +103,16 @@ import Ouroboros.Network.PeerSelection.Governor qualified as Governor
 import Ouroboros.Network.PeerSelection.Governor.Types
            (ChurnMode (ChurnModeNormal), DebugPeerSelection (..),
            PeerSelectionActions, PeerSelectionCounters (..),
-           PeerSelectionPolicy (..), PeerSelectionState,
+           PeerSelectionPolicy (..), PeerSelectionState, TargetsSelector,
            TracePeerSelection (..), emptyPeerSelectionState,
            emptyPublicPeerSelectionState)
 #ifdef POSIX
 import Ouroboros.Network.PeerSelection.Governor.Types
            (makeDebugPeerSelectionState)
 #endif
-import Ouroboros.Network.PeerSelection.LedgerPeers (TraceLedgerPeers,
-           WithLedgerPeersArgs (..))
-import Ouroboros.Network.PeerSelection.LedgerPeers.Type
-           (LedgerPeersConsensusInterface (..), UseLedgerPeers)
+import Ouroboros.Network.PeerSelection.LedgerPeers
+           (LedgerPeersConsensusInterface (..), LedgerStateJudgement (..),
+           TraceLedgerPeers, UseLedgerPeers, WithLedgerPeersArgs (..))
 #ifdef POSIX
 import Ouroboros.Network.PeerSelection.PeerMetric (PeerMetrics,
            fetchynessBlocks, upstreamyness)
@@ -242,10 +241,16 @@ nullTracers =
 data ArgumentsExtra m = ArgumentsExtra {
       -- | selection targets for the peer governor
       --
-      daPeerSelectionTargets :: PeerSelectionTargets
+      daPeerTargetsSelector :: TargetsSelector
 
     , daReadLocalRootPeers  :: STM m [(HotValency, WarmValency, Map RelayAccessPoint (PeerAdvertise, PeerTrustable))]
     , daReadPublicRootPeers :: STM m (Map RelayAccessPoint PeerAdvertise)
+    -- | When syncing up, ie. ledgerStateJudgement == TooOld,
+    -- when this is True we will maintain connection with many big ledger peers
+    -- to get a strong guarantee that when syncing up we will finish with a true
+    -- ledger state. When false, we will fall back on the previous algorithms
+    -- that leverage UseBootstrapPeers flag
+    , daUseGenesis :: Bool
     , daReadUseBootstrapPeers :: STM m UseBootstrapPeers
 
     -- | Peer's own PeerSharing value.
@@ -622,9 +627,10 @@ runM Interfaces
        , daMode = diffusionMode
        }
      ArgumentsExtra
-       { daPeerSelectionTargets
+       { daPeerTargetsSelector
        , daReadLocalRootPeers
        , daReadPublicRootPeers
+       , daUseGenesis
        , daReadUseBootstrapPeers
        , daOwnPeerSharing
        , daReadUseLedgerPeers
@@ -800,12 +806,20 @@ runM Interfaces
 
       churnModeVar <- newTVarIO ChurnModeNormal
 
-      peerSelectionTargetsVar <- newTVarIO $ daPeerSelectionTargets {
+      -- ensure that peer selection governor does not change targets
+      -- in the initial starting phase when churn governor should be
+      -- in control. See comment directly below.
+      churnMutexPeerSelection <- newEmptyTMVarIO
+
+      peerSelectionTargetsVar <- newTVarIO $
+        -- because peer selection governor starts up in TooOld state
+        let base = daPeerTargetsSelector TooOld daUseGenesis
+        in base {
           -- Start with a smaller number of active peers, the churn governor
-          -- will increase it to the configured value after a delay.
+          -- will increase it to the configured value after a delay. This policy
+          -- is maintained in Genesis
           targetNumberOfActivePeers =
-            min 2 (targetNumberOfActivePeers daPeerSelectionTargets)
-        }
+            min 2 (targetNumberOfActivePeers base) }
 
       publicStateVar <- newTVarIO emptyPublicPeerSelectionState
 
@@ -977,7 +991,8 @@ runM Interfaces
                                          psReadUseBootstrapPeers = daReadUseBootstrapPeers,
                                          psPeerSharing = daOwnPeerSharing,
                                          psPeerConnToPeerSharing = pchPeerSharing diNtnPeerSharing,
-                                         psReadPeerSharingController = readTVar (getPeerSharingRegistry daPeerSharingRegistry) }
+                                         psReadPeerSharingController = readTVar (getPeerSharingRegistry daPeerSharingRegistry),
+                                         psChurnMutexPeerSelection = churnMutexPeerSelection }
                                        WithLedgerPeersArgs {
                                          wlpRng = ledgerPeersRng,
                                          wlpConsensusInterface = daLedgerPeersCtx,
@@ -993,15 +1008,16 @@ runM Interfaces
             -> NodeToNodePeerSelectionActions muxMode ntnAddr ntnVersionData m a b
             -> m Void
           peerSelectionGovernor' peerSelectionTracer dbgVar peerSelectionActions =
-              (Governor.peerSelectionGovernor
-                dtTracePeerSelectionTracer
-                peerSelectionTracer
-                dtTracePeerSelectionCounters
-                fuzzRng
-                publicStateVar
-                dbgVar
-                peerSelectionActions
-                peerSelectionPolicy)
+            Governor.peerSelectionGovernor
+              dtTracePeerSelectionTracer
+              peerSelectionTracer
+              dtTracePeerSelectionCounters
+              fuzzRng
+              publicStateVar
+              dbgVar
+              daUseGenesis
+              peerSelectionActions
+              peerSelectionPolicy
 
       --
       -- The peer churn governor:
@@ -1015,9 +1031,11 @@ runM Interfaces
                                  churnModeVar
                                  churnRng
                                  daBlockFetchMode
-                                 daPeerSelectionTargets
+                                 daPeerTargetsSelector
                                  peerSelectionTargetsVar
                                  daReadUseBootstrapPeers
+                                 churnMutexPeerSelection
+                                 daUseGenesis
 
       --
       -- Two functions only used in InitiatorAndResponder mode
@@ -1059,7 +1077,7 @@ runM Interfaces
         -- InitiatorOnly mode, run peer selection only:
         InitiatorOnlyDiffusionMode ->
           withConnectionManagerInitiatorOnlyMode $ \connectionManager-> do
-          debugStateVar <- newTVarIO $ emptyPeerSelectionState fuzzRng []
+          debugStateVar <- newTVarIO $ emptyPeerSelectionState fuzzRng [] daUseGenesis
           diInstallSigUSR1Handler connectionManager debugStateVar daPeerMetrics
           withPeerStateActions' connectionManager $ \peerStateActions->
             withPeerSelectionActions'
@@ -1087,7 +1105,7 @@ runM Interfaces
             inboundInfoChannel
             outboundInfoChannel
             observableStateVar $ \connectionManager-> do
-            debugStateVar <- newTVarIO $ emptyPeerSelectionState fuzzRng []
+            debugStateVar <- newTVarIO $ emptyPeerSelectionState fuzzRng [] daUseGenesis
             diInstallSigUSR1Handler connectionManager debugStateVar daPeerMetrics
             withPeerStateActions' connectionManager $ \peerStateActions ->
               withPeerSelectionActions'